+++ R/utils.R
@@ -11,3 +11,5 @@ is_vector <- function(x) {
 
 "%||%" <- function(a, b) if (is.null(a)) b else a
 
+is_2d <- function(x) length(dim(x)) == 2
+