+++ NEWS.md
@@ -1,3 +1,9 @@
+# dapr 0.0.4
+
+* Functions can now return index names/positions
+* Added row and column-wise versions, i.e., `lapr()`, `vapc_chr()`, etc.
+* Improved documentation
+
 # dapr 0.0.3
 
 * Naming behaviors now mirror defaults of base `lapply()` and `vapply()`, so functions now return names